moved QuestionRow into QuestionCard
