+++ app/components/QuestionCard/QuestionRow/korri.jpg
